Moved all the remaining file actions to the shared folder
