+++ contentcuration/contentcuration/frontend/channelEdit/vuex/contentNode/__tests__/utils.spec.js
@@ -1,7 +1,7 @@
 import each from 'jest-each';
 
 import { ValidationErrors } from '../../../constants';
-import { validateNodeDetails } from '../utils';
+import { validateNodeDetails, validateNodeFiles } from '../utils';
 
 describe('utils', () => {
   describe('validateNodeDetails', () => {
@@ -159,6 +159,54 @@ describe('utils', () => {
       ],
     ]).it('validates mastery model for exercises', (node, errors) => {
       expect(validateNodeDetails(node)).toEqual(errors);
+    });
+  });
+
+  describe('validateNodeFiles', () => {
+    it('throws an error if there are no valid primary files', () => {
+      let testFiles = [
+        {
+          error: ValidationErrors.UPLOAD_FAILED,
+          preset: { supplementary: false },
+        },
+        {
+          error: ValidationErrors.UPLOAD_FAILED,
+          preset: { supplementary: false },
+        },
+        {
+          preset: { supplementary: true },
+        },
+      ];
+      expect(validateNodeFiles(testFiles)).toContain(ValidationErrors.NO_VALID_PRIMARY_FILES);
+    });
+    it('does not throw NO_VALID_PRIMARY_FILES if there is one valid primary file', () => {
+      let testFiles = [
+        {
+          error: ValidationErrors.UPLOAD_FAILED,
+          preset: { supplementary: false },
+        },
+        {
+          preset: { supplementary: false },
+        },
+      ];
+      expect(validateNodeFiles(testFiles)).not.toContain(ValidationErrors.NO_VALID_PRIMARY_FILES);
+    });
+    it('returns array of errors found on files', () => {
+      let testFiles = [
+        {
+          error: ValidationErrors.NO_STORAGE,
+          preset: { supplementary: true },
+        },
+        {
+          error: ValidationErrors.UPLOAD_FAILED,
+          preset: { supplementary: true },
+        },
+        {
+          preset: { supplementary: false },
+        },
+      ];
+      let expectedErrors = [ValidationErrors.NO_STORAGE, ValidationErrors.UPLOAD_FAILED];
+      expect(validateNodeFiles(testFiles)).toEqual(expectedErrors);
     });
   });
 });
