+++ contentcuration/contentcuration/frontend/channelEdit/vuex/contentNode/utils.js
@@ -46,3 +46,18 @@ export function validateNodeDetails(node) {
 
   return errors;
 }
+
+/**
+ * Validate node files - correct types, no associated errors, etc.
+ * @param {Array} files An array of files for a node.
+ * @returns {Array} An array of error codes.
+ */
+export function validateNodeFiles(files) {
+  let errors = files.filter(f => f.error).map(f => f.error);
+  let validPrimaryFiles = files.filter(f => !f.error && !f.preset.supplementary);
+
+  if (!validPrimaryFiles.length) {
+    errors.push(ValidationErrors.NO_VALID_PRIMARY_FILES);
+  }
+  return errors;
+}
