+++ contentcuration/contentcuration/frontend/shared/vuex/baseStore.js
@@ -5,10 +5,8 @@ import ConnectionPlugin from './connectionPlugin';
 import snackbar from './snackbar';
 import contextMenu from './contextMenu';
 import channel from './channel';
+import file from './file';
 import policies from './policies';
-
-// TODO: once direct file uploading is done, move that to this folder
-import file from 'frontend/channelEdit/vuex/file';
 import { registerListener } from 'shared/data';
 
 Vue.use(Vuex);
