+++ contentcuration/contentcuration/frontend/shared/vuex/file/actions.js
@@ -180,6 +180,7 @@ export function uploadFile(context, { file }) {
             return context
               .dispatch('uploadFileToStorage', { checksum, file, url: response.data })
               .then(response => {
+                response.data = `/content/storage/${checksum[0]}/${checksum[1]}/${checksum}.jpg`;
                 context.commit('ADD_FILEUPLOAD', { checksum, file_on_disk: response.data });
               })
               .catch(() => {
